// src/lib/types/group_type.ts
export interface GroupType {
    GroupNo:          string;
    GroupName:        string;
    GroupCode:        string;
    status:           string;
    RegistrationDate: string;
    RegCertNo:        string;
    PhoneNumber:      string;
    PostalAddress:    string;
    PostalCode:       string;
    Town:             string;
    GroupDescription: string;
    DateCreated:      Date;
    ConstituencyName: string;
    CountyName:       string;
    WardName:         string;
}